finished gif-bg

+++ tarot/sketch.js
@@ -1,49 +1,43 @@
-let card
-//let s = 0;
-// let w = 10;
-// let h = 17;
-//let zoom = 0;
+let card;
 let cards;
+let last;
 
-function setup(){
-    createCanvas(windowWidth, windowHeight);
 
-card = loadImage('deck/turned.png')
-cards = Array().fill().map(c =>new Card())
 
+function setup() {
+  createCanvas(130,220) ;
+  card = loadImage("deck/turned.png");
+  cards = Array().fill().map(c => new Card());
+  frameRate(10)
+    // createLoop({duration:6, gif:true,startLoop:3, endLoop:3})
+   
+}
+
+function draw() {
+
+  //background(0);
+  translate(width / 2, height / 2);
+
+  for (let i = 0; i < 1; i++) {
+    let c = new Card();
+    cards.push(c);
+  }
+
+  for (let card of cards) { 
+    card.display();
+    card.update();
+
+  }
+
+
+   if(cards.length > 100){
+    cards = cards.filter(c => !c.delete());
+    //cards=[]
+   
+   }
+
+   
 
 }
 
-function draw(){
-    background(0)
-    translate(width/2,height/2)
-for(let i=0; i<1; i++){
 
-let c = new Card()
-cards.push(c)
-
-}
-
-for(let card of cards){
-    card.display()
-    card.update()
-}
-cards = cards.filter(c => !c.delete())
-
-//console.log(cards);
-
-//   if(zoom >= 90){
-//       zoom=0
-//       w=10
-//       h=17
-//   }
-
-
-//tint(255,100)
-// scale(zoom,zoom)
-
-// image(card, 0,0,w,h)
-
-// zoom+=0.5
-
-}
